Sketches and Libraries added
Two processing sketches added in Mind_process_STL_Max folder. The main sketch with a longer name is the sketch that will be used in the end. The other sketch generates pyramids, this is for Ali to experiment around the pyramid's appearance. The changes maybe merged later to the original sketch.

Libraries folder should contain all the libraries needed to run either of the sketches.